types: D_800DEF90 is an array of void (*)(GObj *), not void (*)(s32)
Two lanes independently traced conflict families back to this one
declaration, so it is fixed at the root rather than worked around 98 more
times.

The array was declared `extern void (*D_800DEF90[])(s32);` in four places
while every callback installed into it -- func_800B6A2C, func_800B6B8C,
func_800B4924, func_800B7514, func_800B6E84 -- is defined `void (GObj *)`.
The dispatch site proved it: ovl1_7.c's func_800B0D90 called it as

    D_800DEF90[omCurrentObj->objId]((s32) arg0);

casting a GObj * to s32 to satisfy the wrong declaration. That is a live LP64
truncation in the PC port, not a cosmetic disagreement -- a host pointer does
not survive the round trip through s32.

With the array correctly typed, the workarounds it forced all disappear:

  - 90 of the 98 `(void (*)(s32))` casts at assignment sites wrapped a callee
    that already takes GObj *, so they were pure noise and are gone.
  - The other 8 wrapped func_800B531C, defined `void (s32)` in ovl1_8.c while
    all its callers pass a GObj *. Its argument is unused -- the PORT arm even
    writes `(void) arg0;` -- so correcting the parameter is semantically free,
    and it retires the last 8 casts and 2 stale declarations.
  - The `void (s32)` declarations in ovl9_1/3/7 existed only so the assignment
    would type-check, and the explicit casts in ovl9_7 were the same
    workaround written out.

track_arrays.h had already recorded the suspicion that the real element type
was `void (*)(GObj *)`; this confirms it from the definitions and the call.

ROM sha1 6cea2d46b929 unchanged across all 25 files; PC build 0 errors.

+++ src/ovl1/ovl1_7.c
@@ -737,7 +737,7 @@ void func_800B0D90(GObj *arg0) {
         }
         if (!(D_800DD8D0[omCurrentObj->objId] & 4)) {
             if (D_800DEF90[omCurrentObj->objId] != NULL) {
-                D_800DEF90[omCurrentObj->objId]((s32) arg0);
+                D_800DEF90[omCurrentObj->objId](arg0);
             }
         }
         D_800E56D0[omCurrentObj->objId] = 0.0f;
